feat: consolidate Promotion construction in defaulting webhook
API server endpoints and the Stage controller now submit minimal
Promotions (Stage + Freight); the defaulting webhook unconditionally
populates the name, steps, and vars from the referenced Stage's
PromotionTemplate. As a result, kubectl-driven Promotion creation
now behaves the same as creation via the UI or CLI.

Closes #5126.
