+++ pkg/webhook/kubernetes/promotion/webhook.go
@@ -22,7 +22,6 @@ import (
 	"github.com/akuity/kargo/pkg/api"
 	kargoEvent "github.com/akuity/kargo/pkg/event"
 	k8sevent "github.com/akuity/kargo/pkg/event/kubernetes"
-	"github.com/akuity/kargo/pkg/kargo"
 	libEvent "github.com/akuity/kargo/pkg/kubernetes/event"
 	"github.com/akuity/kargo/pkg/logging"
 	libWebhook "github.com/akuity/kargo/pkg/webhook/kubernetes"
@@ -134,44 +133,9 @@ func (w *webhook) Default(ctx context.Context, obj runtime.Object) error {
 	}
 
 	promo := obj.(*kargoapi.Promotion) // nolint: forcetypeassert
-	var oldPromo *kargoapi.Promotion
-	// We need to decode old object manually since controller-runtime doesn't decode it for us.
-	if req.Operation == admissionv1.Update {
-		oldPromo = &kargoapi.Promotion{}
-		if err = w.decoder.DecodeRaw(req.OldObject, oldPromo); err != nil {
-			return fmt.Errorf("decode old object: %w", err)
-		}
-	}
 
 	if promo.Annotations == nil {
 		promo.Annotations = make(map[string]string, 1)
-	}
-
-	switch req.Operation {
-	case admissionv1.Create:
-		// Set actor as an admission request's user info when the promotion is created
-		// to allow controllers to track who created it.
-		if !w.isRequestFromKargoControlplaneFn(req) {
-			promo.Annotations[kargoapi.AnnotationKeyCreateActor] = api.FormatEventKubernetesUserActor(req.UserInfo)
-		}
-
-		// Enrich the annotation with the actor and control plane information.
-		w.setAbortAnnotationActor(req, nil, promo)
-
-		// Inflate any PromotionTasks in the Promotion's steps
-		if err = kargo.NewPromotionBuilder(w.client).InflateSteps(ctx, promo); err != nil {
-			return fmt.Errorf("failed to inflate Promotion steps: %w", err)
-		}
-	case admissionv1.Update:
-		// Ensure actor annotation immutability
-		if oldActor, ok := oldPromo.Annotations[kargoapi.AnnotationKeyCreateActor]; ok {
-			promo.Annotations[kargoapi.AnnotationKeyCreateActor] = oldActor
-		} else {
-			delete(promo.Annotations, kargoapi.AnnotationKeyCreateActor)
-		}
-
-		// Enrich the annotation with the actor and control plane information.
-		w.setAbortAnnotationActor(req, oldPromo, promo)
 	}
 
 	stage, err := w.getStageFn(
@@ -197,16 +161,75 @@ func (w *webhook) Default(ctx context.Context, obj runtime.Object) error {
 			promo.Namespace,
 		)
 	}
-	if len(promo.Spec.Steps) == 0 {
-		// nolint:staticcheck
-		return fmt.Errorf(
-			"Stage %q in namespace %q defines no promotion steps",
-			promo.Spec.Stage,
-			promo.Namespace,
+
+	switch req.Operation {
+	case admissionv1.Create:
+		// Comparison and sorting logic elsewhere in the Kargo code base depend on
+		// adhering to a specific naming convention. Always overwrite the name with
+		// one generated by Kargo.
+		promo.Name = api.GeneratePromotionName(stage.Name, promo.Spec.Freight)
+
+		// Set actor as an admission request's user info when the promotion is
+		// created to allow controllers to track who created it.
+		if !w.isRequestFromKargoControlplaneFn(req) {
+			promo.Annotations[kargoapi.AnnotationKeyCreateActor] = api.FormatEventKubernetesUserActor(req.UserInfo)
+		}
+
+		// Enrich the annotation with the actor and control plane information.
+		w.setAbortAnnotationActor(req, nil, promo)
+
+		// Reject the Promotion if the Stage's PromotionTemplate is missing or
+		// defines no steps. Done here (in the mutating webhook) for the sake of
+		// surfacing a nicer, more informative error message than the one that one
+		// would otherwise get when the CRD's declarative validation fails due to
+		// the missing steps.
+		if stage.Spec.PromotionTemplate == nil || len(stage.Spec.PromotionTemplate.Spec.Steps) == 0 {
+			// nolint:staticcheck
+			return fmt.Errorf(
+				"Stage %q in namespace %q defines no promotion steps",
+				promo.Spec.Stage,
+				promo.Namespace,
+			)
+		}
+
+		// Promotions must always follow the process defined by the Stage they
+		// reference. Deviation from that is not supported. Unconditionally
+		// (re-)populate the Promotion's steps and vars according to the Stage's
+		// PromotionTemplate.
+		promo.Spec.Steps = stage.Spec.PromotionTemplate.Spec.Steps
+		vars := make(
+			[]kargoapi.ExpressionVariable,
+			0,
+			len(stage.Spec.Vars)+len(stage.Spec.PromotionTemplate.Spec.Vars),
 		)
+		vars = append(vars, stage.Spec.Vars...)
+		vars = append(vars, stage.Spec.PromotionTemplate.Spec.Vars...)
+		promo.Spec.Vars = vars
+
+		// Inflate any PromotionTasks in the Promotion's steps.
+		if err = api.InflateSteps(ctx, w.client, promo); err != nil {
+			return fmt.Errorf("failed to inflate Promotion steps: %w", err)
+		}
+	case admissionv1.Update:
+		// We need to decode the old object manually since controller-runtime
+		// doesn't decode it for us.
+		oldPromo := &kargoapi.Promotion{}
+		if err = w.decoder.DecodeRaw(req.OldObject, oldPromo); err != nil {
+			return fmt.Errorf("decode old object: %w", err)
+		}
+
+		// Ensure actor annotation immutability.
+		if oldActor, ok := oldPromo.Annotations[kargoapi.AnnotationKeyCreateActor]; ok {
+			promo.Annotations[kargoapi.AnnotationKeyCreateActor] = oldActor
+		} else {
+			delete(promo.Annotations, kargoapi.AnnotationKeyCreateActor)
+		}
+
+		// Enrich the annotation with the actor and control plane information.
+		w.setAbortAnnotationActor(req, oldPromo, promo)
 	}
 
-	// Make sure the Promotion has the same shard as the Stage
+	// Make sure the Promotion has the same shard as the Stage.
 	if stage.Spec.Shard != "" {
 		if promo.Labels == nil {
 			promo.Labels = make(map[string]string, 1)
